+++ src/python_api.rs
@@ -84,8 +84,8 @@ use crate::connectors::data_storage::{
     new_csv_filesystem_reader, new_filesystem_reader, new_s3_csv_reader, new_s3_generic_reader,
     ConnectorMode, DeltaTableReader, ElasticSearchWriter, FileWriter, IcebergReader, KafkaReader,
     KafkaWriter, LakeWriter, MessageQueueTopic, MongoWriter, NatsReader, NatsWriter, NullWriter,
-    ObjectDownloader, PsqlWriter, PythonConnectorEventType, PythonReaderBuilder, ReadError,
-    ReadMethod, ReaderBuilder, SqlWriterInitMode, SqliteReader, WriteError, Writer,
+    ObjectDownloader, PsqlWriter, PythonConnectorEventType, PythonReaderBuilder, RdkafkaWatermark,
+    ReadError, ReadMethod, ReaderBuilder, SqlWriterInitMode, SqliteReader, WriteError, Writer,
 };
 use crate::connectors::scanner::S3Scanner;
 use crate::connectors::synchronization::ConnectorGroupDescriptor;
@@ -4808,15 +4808,94 @@ impl DataStorage {
         Ok((Box::new(reader), 1))
     }
 
+    /// Returns the total number of partitions for a Kafka topic
     fn total_partitions_for_topic(consumer: &BaseConsumer, topic: &str) -> PyResult<usize> {
         let metadata = consumer
-            .fetch_metadata(Some(topic), std::time::Duration::from_secs(30))
+            .fetch_metadata(Some(topic), KafkaReader::default_timeout())
             .map_err(|e| PyIOError::new_err(format!("Failed to fetch topic metadata: {e}")))?;
         if let Some(topic) = metadata.topics().iter().find(|t| t.name() == topic) {
             Ok(topic.partitions().len())
         } else {
             Err(PyIOError::new_err(format!("Topic '{topic}' not found")))
         }
+    }
+
+    /// Returns an array of partition watermarks.
+    /// Used to handle cases where a later call to `offsets_for_times`
+    /// might return `KafkaOffset::End` for some partitions, allowing for graceful handling.
+    /// Also used in static mode to identify the boundaries of the data chunk that needs to be read.
+    fn kafka_partition_watermarks(
+        consumer: &BaseConsumer,
+        topic: &str,
+        total_partitions: usize,
+    ) -> PyResult<Vec<RdkafkaWatermark>> {
+        let mut next_used_offset_per_partition = Vec::with_capacity(total_partitions);
+        for partition_idx in 0..total_partitions {
+            let (start_offset, next_offset) = consumer
+                .fetch_watermarks(
+                    topic,
+                    partition_idx.try_into().unwrap(),
+                    KafkaReader::default_timeout(),
+                )
+                .map_err(|e| {
+                    PyIOError::new_err(format!(
+                        "Failed to fetch watermarks for ({topic}, {partition_idx}): {e}"
+                    ))
+                })?;
+            next_used_offset_per_partition.push(RdkafkaWatermark::new(start_offset, next_offset));
+        }
+        Ok(next_used_offset_per_partition)
+    }
+
+    fn kafka_seek_positions_for_timestamp(
+        consumer: &BaseConsumer,
+        topic: &str,
+        total_partitions: usize,
+        start_from_timestamp_ms: i64,
+        watermarks: &[RdkafkaWatermark],
+    ) -> PyResult<HashMap<i32, KafkaOffset>> {
+        let mut seek_positions = HashMap::new();
+        let mut tpl = TopicPartitionList::new();
+        for partition_idx in 0..total_partitions {
+            tpl.add_partition_offset(
+                topic,
+                partition_idx.try_into().unwrap(),
+                KafkaOffset::Offset(start_from_timestamp_ms),
+            )
+            .expect("Failed to add partition offset");
+        }
+
+        let offsets = consumer
+            .offsets_for_times(tpl, KafkaReader::default_timeout())
+            .map_err(|e| {
+                PyIOError::new_err(format!("Failed to fetch offsets for the timestamp: {e}"))
+            })?;
+
+        // We could have done a simple `consumer.assign` here, but it would damage the automatic consumer rebalance
+        // So we act differently: we pass the seek positions to consumer, and it seeks lazily
+        for element in offsets.elements() {
+            assert_eq!(element.topic(), topic);
+            let offset = match element.offset() {
+                KafkaOffset::Invalid => {
+                    return Err(PyRuntimeError::new_err(format!(
+                        "rdkafka returned invalid offset, details: {offsets:?}"
+                    )))
+                }
+                KafkaOffset::End => {
+                    let partition_idx: usize = element.partition().try_into().unwrap();
+                    info!("Partition {partition_idx} doesn't have messages with timestamp greater than {start_from_timestamp_ms}. All new messages will be read.");
+                    KafkaOffset::Offset(watermarks[partition_idx].high)
+                }
+                offset => offset,
+            };
+            info!(
+                "Adding a lazy seek position for ({topic}, {}) to ({:?})",
+                element.partition(),
+                offset
+            );
+            seek_positions.insert(element.partition(), offset);
+        }
+        Ok(seek_positions)
     }
 
     fn construct_kafka_reader(&self) -> PyResult<(Box<dyn ReaderBuilder>, usize)> {
@@ -4831,75 +4910,30 @@ impl DataStorage {
             .subscribe(&[topic])
             .map_err(|e| PyIOError::new_err(format!("Subscription to Kafka topic failed: {e}")))?;
 
+        let total_partitions = Self::total_partitions_for_topic(&consumer, topic)?;
+        let watermarks = Self::kafka_partition_watermarks(&consumer, topic, total_partitions)?;
+
         let mut seek_positions = HashMap::new();
         if let Some(start_from_timestamp_ms) = self.start_from_timestamp_ms {
             let current_timestamp = current_unix_timestamp_ms();
             if start_from_timestamp_ms > current_timestamp.try_into().unwrap() {
                 warn!("The timestamp {start_from_timestamp_ms} is greater than the current timestamp {current_timestamp}. All new entries will be read.");
             }
-
-            let mut tpl = TopicPartitionList::new();
-            let total_partitions = Self::total_partitions_for_topic(&consumer, topic)?;
-            for partition_idx in 0..total_partitions {
-                tpl.add_partition_offset(
-                    topic,
-                    partition_idx.try_into().unwrap(),
-                    KafkaOffset::Offset(start_from_timestamp_ms),
-                )
-                .expect("Failed to add partition offset");
-            }
-
-            // First we find the next offset within partition to gracefully handle the case
-            // when the later call of `offsets_for_times` returns `KafkaOffset::End` for some
-            // of the partitions
-            let mut next_used_offset_per_partition = Vec::with_capacity(total_partitions);
-            for partition_idx in 0..total_partitions {
-                let (_, next_offset) = consumer
-                    .fetch_watermarks(
-                        topic,
-                        partition_idx.try_into().unwrap(),
-                        std::time::Duration::from_secs(30),
-                    )
-                    .map_err(|e| {
-                        PyIOError::new_err(format!(
-                            "Failed to fetch watermarks for ({topic}, {partition_idx}): {e}"
-                        ))
-                    })?;
-                next_used_offset_per_partition.push(KafkaOffset::Offset(next_offset));
-            }
-
-            let offsets = consumer
-                .offsets_for_times(tpl, std::time::Duration::from_secs(30))
-                .map_err(|e| {
-                    PyIOError::new_err(format!("Failed to fetch offsets for the timestamp: {e}"))
-                })?;
-
-            // We could have done a simple `consumer.assign` here, but it would damage the automatic consumer rebalance
-            // So we act differently: we pass the seek positions to consumer, and it seeks lazily
-            for element in offsets.elements() {
-                assert_eq!(element.topic(), topic);
-                let offset = match element.offset() {
-                    KafkaOffset::Invalid => {
-                        return Err(PyRuntimeError::new_err(format!(
-                            "rdkafka returned invalid offset, details: {offsets:?}"
-                        )))
-                    }
-                    KafkaOffset::End => {
-                        let partition_idx: usize = element.partition().try_into().unwrap();
-                        info!("Partition {partition_idx} doesn't have messages with timestamp greater than {start_from_timestamp_ms}. All new messages will be read.");
-                        next_used_offset_per_partition[partition_idx]
-                    }
-                    offset => offset,
-                };
-                info!(
-                    "Adding a lazy seek position for ({topic}, {}) to ({:?})",
-                    element.partition(),
-                    offset
-                );
-                seek_positions.insert(element.partition(), offset);
-            }
+            seek_positions = Self::kafka_seek_positions_for_timestamp(
+                &consumer,
+                topic,
+                total_partitions,
+                start_from_timestamp_ms,
+                &watermarks,
+            )?;
         }
-        let reader = KafkaReader::new(consumer, topic.to_string(), seek_positions);
+        let reader = KafkaReader::new(
+            consumer,
+            topic.to_string(),
+            seek_positions,
+            watermarks,
+            self.mode,
+        );
         Ok((Box::new(reader), self.parallel_readers.unwrap_or(256)))
     }